:recycle: move RequestBodyCaptureMethod to undertow-specific library
