+++ instrumentation/undertow/undertow-1.4/build.gradle.kts
@@ -27,6 +27,7 @@ val versions: Map<String, String> by extra
 dependencies {
     implementation("io.opentelemetry.javaagent.instrumentation:opentelemetry-javaagent-undertow-1.4:${versions["opentelemetry_java_agent"]}")
     library("io.undertow:undertow-core:2.0.0.Final")
+    implementation(project(":instrumentation:undertow:undertow-common"))
     testImplementation(testFixtures(project(":testing-common")))
     testImplementation("javax.servlet:javax.servlet-api:3.1.0")
     testImplementation("io.undertow:undertow-servlet:2.0.0.Final")
