+++ instrumentation/undertow/undertow-common/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    `java-library`
+}
