+++ instrumentation/undertow/undertow-servlet-1.4/build.gradle.kts
@@ -24,4 +24,5 @@ afterEvaluate{
 
 dependencies {
     library("io.undertow:undertow-servlet:2.0.0.Final")
+    implementation(project(":instrumentation:undertow:undertow-common"))
 }
