+++ settings.gradle.kts
@@ -52,6 +52,8 @@ include("instrumentation:servlet:servlet-rw")
 findProject(":instrumentation:servlet:servlet-rw")?.name = "servlet-rw"
 include("instrumentation:struts-2.3")
 findProject(":instrumentation:struts-2.3")?.name = "struts-2.3"
+include("instrumentation:undertow:undertow-common")
+findProject(":instrumentation:undertow:undertow-common")?.name = "undertow-common"
 include("instrumentation:undertow:undertow-1.4")
 findProject(":instrumentation:undertow:undertow-1.4")?.name = "undertow-1.4"
 include("instrumentation:undertow:undertow-servlet-1.4")
